Added minor changes to the app
Replaced push with push replacement and changed oversight of using int instead of double.

+++ lib/splash_screen.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:mobile_banking_app/authentication_screen.dart';
 import 'package:flutter/material.dart';
 
-List<int> balance = [75000, 25000, 0]; 
+List<double> balance = [75000, 25000, 0]; 
 
 class SplashScreen extends StatefulWidget {
   const SplashScreen({Key? key}) : super(key: key);
